requirements.txt
aiohttp>=3.8.0
datasets>=2.14.0
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
tqdm>=4.65.0
huggingface_hub>=0.16.0